# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API documentation for Bonpay Finance API",
        "title": "Bonpay Finance",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Johnson Awah Alfred",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://api.bonpay.finance",
    "basePath": "/api/v1",
    "paths": {
...
        "/user": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates some details about the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "update user profile",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateUserProfileInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile information of the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.UpdateUserProfileInput": {
            "type": "object",
            "required": [
                "buisness_name",
                "company_logo",
                "company_website",
                "manager",
                "phone_number"
            ],
            "properties": {
                "buisness_name": {
                    "type": "string"
                },
                "company_logo": {
                    "type": "string"
                },
                "company_website": {
                    "type": "string"
                },
                "manager": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "handlers.User": {
            "type": "object",
            "required": [
                "email_verified",
                "full_name",
                "user_id"
            ],
            "properties": {
                "account_verified": {
                    "type": "boolean"
                },
                "business_industry": {
                    "type": "string"
                },
                "business_website": {
                    "type": "string"
                },
                "company_logo_url": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "is_software_developer": {
                    "type": "boolean"
                },
                "last_login": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "wallet": {
                    "$ref": "#/definitions/models.Wallet"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Wallet": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pk": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sol_balance_devnet": {
                    "type": "number"
                },
                "sol_balance_mainnet": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "usdc_balance_devnnet": {
                    "type": "number"
                },
                "usdc_balance_mainnet": {
                    "type": "number"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        }
    }
}